+++ frontend/src/components/index.tsx
@@ -3,3 +3,4 @@ export * from './common';
 export * from './error';
 export * from './highlight/components';
 export * from './login';
+export * from './reviewURL';
