+++ frontend/src/components/reviewURL/PasswordField/style.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const PasswordInputContainer = styled.div`
+  position: relative;
+  display: flex;
+`;
