+++ frontend/src/components/reviewURL/index.ts
@@ -2,6 +2,5 @@ export { default as FormBody } from './FormBody';
 export { default as FormLayout } from './FormLayout';
 export { default as ReviewZoneURLModal } from './ReviewZoneURLModal';
 export { default as URLGeneratorForm } from './URLGeneratorForm';
-export { default as CopyTextButton } from './CopyTextButton';
-export { default as ReviewMeOverview } from './ReviewMeOverview';
-export { default as InfinityCarousel } from './InfinityCarousel';
+export { default as PasswordField } from './PasswordField';
+export { default as ReviewGroupDataField } from './ReviewGroupDataField';
