--- frontend/src/pages/HomePage/components/Inputs/index.ts
@@ -1,3 +1,0 @@
-export { default as ReviewGroupDataField } from './ReviewGroupDataField';
-export { default as PasswordField } from './PasswordField';
-export { default as InputField } from './InputField';
